add message object to abstract mail gem interactions
Mailthis is essentially only using the mail gem for it's message
API and for building the message string to send using `Net::SMTP`.
The previous implementation used the `::Mail` object directly.  This
switches to abstracting the mail interactions via a message class.

We really only need a small subset of mails capabilities and this
enforces that.  It also establishes a spec for what a mailthis
message needs to do.  Finally, this implementation more easily allows
for switching out the mail gem for a better alternative when/if
one arises.

Note: this also cleans up some of things about the outgoing email's
implementation.  These are mostly just cleanups and attempts at
keeping the implementation easy to follow and comprehend.

Overall, no behavior changes.