add test mailer for use in tests
This mailer tracks delivered messages so you can test that your
code delivers messages appropriately.

To do this, I followed the pattern from Sanford/Deas/Qs/etc and
made `Mailer` a module that returns either a `MailthisMailer` or a
`TestMailer` depending on if in test mode or not.  Test mode is
driven by the `MAILTHIS_TEST_MODE` env var.