remove ns-options as a dependency
This is part of getting this gem working in modern ruby versions.
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This switches the mailer's settings to be set manually by a DSL
method for each setting.  This includes honoring the previous
defaulting logic for `smtp_auth`, `from` and `logger`.  This also
manually implements the `valid?` and  `validate!` logic.  Note:
there are some backwards-incompatible changes here, specifically
you can't no longer build mailers given a hash of options and
you can't set option values to procs for lazy-eval.  These were
features previously provided by ns-options.  You also can't eet
options to `nil` values like you could before.

I did a bunch of reworks to the mailer tests, mostly related to
not being able to set individual settings to `nil` and the
manual defaulting logic.  This involved updating the mailer
factory to take a hash of settings whiched helped in creating
the necessary mailers for these tests.  A few other tests were
modified as needed as well.

Finally, this brings in much-plugin to ensure the mailer mixin
logic is only applied once.